+++ registry/service_test.go
@@ -1,4 +1,4 @@
-package cascade
+package registry
 
 import (
 	"sync"
@@ -35,7 +35,7 @@ func TestRepository(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		service := NewRegistryService(tt.constructor())
+		service := NewService(tt.constructor())
 
 		t.Run(tt.description, func(t *testing.T) {
 			t.Run("Create a repository, do something with it, and delete it", func(t *testing.T) {
