+++ testing/conformance/pull_test.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/robinkb/cascade-registry"
 	v2 "github.com/robinkb/cascade-registry/api/v2"
+	"github.com/robinkb/cascade-registry/registry"
 	"github.com/robinkb/cascade-registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
 	testclient "github.com/robinkb/cascade-registry/testing/client"
@@ -15,7 +15,7 @@ import (
 func TestPull(t *testing.T) {
 	metadata := inmemory.NewMetadataStore()
 	blobs := inmemory.NewBlobStore()
-	service := cascade.NewRegistryService(metadata, blobs)
+	service := registry.NewService(metadata, blobs)
 	srv := v2.New(service)
 
 	t.Run("Pulling manifests", func(t *testing.T) {
